최종 수정

+++ src/main/java/com/totoblog/data/repository/KeywordRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface KeywordRepository extends JpaRepository<Keyword, Long> {
+public interface KeywordRepository extends JpaRepository<Keyword, Integer> {
 
 
     /*** 인기 검색어 10개 조회 ***/
